[Refactor] Refactor Delete Protected Instances
