+++ openstack/sdrs/v1/protectedinstances/requests.go
@@ -5,38 +5,6 @@ import (
 	"github.com/opentelekomcloud/gophertelekomcloud/pagination"
 )
 
-// DeleteOptsBuilder allows extensions to add additional parameters to the
-// Delete request.
-type DeleteOptsBuilder interface {
-	ToInstanceDeleteMap() (map[string]interface{}, error)
-}
-
-// DeleteOpts contains all the values needed to delete an Instance.
-type DeleteOpts struct {
-	// Delete Target Server
-	DeleteTargetServer *bool `json:"delete_target_server,omitempty"`
-	// Delete Target Eip
-	DeleteTargetEip *bool `json:"delete_target_eip,omitempty"`
-}
-
-// ToInstanceDeleteMap builds a update request body from DeleteOpts.
-func (opts DeleteOpts) ToInstanceDeleteMap() (map[string]interface{}, error) {
-	return golangsdk.BuildRequestBody(opts, "")
-}
-
-// Delete will permanently delete a particular Instance based on its unique ID.
-func Delete(c *golangsdk.ServiceClient, id string, opts DeleteOptsBuilder) (r JobResult) {
-	b, err := opts.ToInstanceDeleteMap()
-	if err != nil {
-		r.Err = err
-		return
-	}
-	_, r.Err = c.DeleteWithBodyResp(resourceURL(c, id), b, &r.Body, &golangsdk.RequestOpts{
-		OkCodes: []int{200},
-	})
-	return
-}
-
 type ListOptsBuilder interface {
 	ToInstanceListQuery() (string, error)
 }
